refactor: shop layout server component 정리
